+++ lib/broker/try_test.go
@@ -1,0 +1,316 @@
+// Copyright (c) 2018 Ashley Jeffs
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package broker
+
+import (
+	"errors"
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/Jeffail/benthos/lib/metrics"
+	"github.com/Jeffail/benthos/lib/types"
+)
+
+//------------------------------------------------------------------------------
+
+func TestTryInterfaces(t *testing.T) {
+	f := &Try{}
+	if types.Consumer(f) == nil {
+		t.Errorf("Try: nil types.Consumer")
+	}
+	if types.Closable(f) == nil {
+		t.Errorf("Try: nil types.Closable")
+	}
+}
+
+func TestTryDoubleClose(t *testing.T) {
+	oTM, err := NewTry([]types.Output{}, metrics.DudType{})
+	if err != nil {
+		t.Error(err)
+		return
+	}
+
+	// This shouldn't cause a panic
+	oTM.CloseAsync()
+	oTM.CloseAsync()
+}
+
+//------------------------------------------------------------------------------
+
+func TestTryHappyPath(t *testing.T) {
+	outputs := []types.Output{}
+	mockOutputs := []*MockOutputType{
+		{},
+		{},
+		{},
+	}
+
+	for _, o := range mockOutputs {
+		outputs = append(outputs, o)
+	}
+
+	readChan := make(chan types.Transaction)
+	resChan := make(chan types.Response)
+
+	oTM, err := NewTry(outputs, metrics.DudType{})
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	if err = oTM.StartReceiving(readChan); err != nil {
+		t.Error(err)
+		return
+	}
+
+	for i := 0; i < 10; i++ {
+		content := [][]byte{[]byte(fmt.Sprintf("hello world %v", i))}
+		select {
+		case readChan <- types.NewTransaction(types.NewMessage(content), resChan):
+		case <-time.After(time.Second):
+			t.Errorf("Timed out waiting for broker send")
+			return
+		}
+
+		go func() {
+			var ts types.Transaction
+			select {
+			case ts = <-mockOutputs[0].TChan:
+				if string(ts.Payload.Get(0)) != string(content[0]) {
+					t.Errorf("Wrong content returned %s != %s", ts.Payload.Get(0), content[0])
+				}
+			case <-mockOutputs[1].TChan:
+				t.Error("Received message in wrong order")
+				return
+			case <-mockOutputs[2].TChan:
+				t.Error("Received message in wrong order")
+				return
+			case <-time.After(time.Second):
+				t.Errorf("Timed out waiting for broker propagate")
+				return
+			}
+
+			select {
+			case ts.ResponseChan <- types.NewSimpleResponse(nil):
+			case <-time.After(time.Second):
+				t.Errorf("Timed out responding to broker")
+				return
+			}
+		}()
+
+		select {
+		case res := <-resChan:
+			if res.Error() != nil {
+				t.Error(res.Error())
+			}
+		case <-time.After(time.Second):
+			t.Errorf("Timed out responding to broker")
+			return
+		}
+	}
+
+	oTM.CloseAsync()
+	if err := oTM.WaitForClose(time.Second * 10); err != nil {
+		t.Error(err)
+	}
+}
+
+func TestTryHappyishPath(t *testing.T) {
+	outputs := []types.Output{}
+	mockOutputs := []*MockOutputType{
+		{},
+		{},
+		{},
+	}
+
+	for _, o := range mockOutputs {
+		outputs = append(outputs, o)
+	}
+
+	readChan := make(chan types.Transaction)
+	resChan := make(chan types.Response)
+
+	oTM, err := NewTry(outputs, metrics.DudType{})
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	if err = oTM.StartReceiving(readChan); err != nil {
+		t.Error(err)
+		return
+	}
+
+	for i := 0; i < 10; i++ {
+		content := [][]byte{[]byte(fmt.Sprintf("hello world %v", i))}
+		select {
+		case readChan <- types.NewTransaction(types.NewMessage(content), resChan):
+		case <-time.After(time.Second):
+			t.Errorf("Timed out waiting for broker send")
+			return
+		}
+
+		go func() {
+			var ts types.Transaction
+			select {
+			case ts = <-mockOutputs[0].TChan:
+				if string(ts.Payload.Get(0)) != string(content[0]) {
+					t.Errorf("Wrong content returned %s != %s", ts.Payload.Get(0), content[0])
+				}
+			case <-mockOutputs[1].TChan:
+				t.Error("Received message in wrong order")
+				return
+			case <-mockOutputs[2].TChan:
+				t.Error("Received message in wrong order")
+				return
+			case <-time.After(time.Second):
+				t.Errorf("Timed out waiting for broker propagate")
+				return
+			}
+
+			select {
+			case ts.ResponseChan <- types.NewSimpleResponse(errors.New("test err")):
+			case <-time.After(time.Second):
+				t.Errorf("Timed out responding to broker")
+				return
+			}
+
+			select {
+			case ts = <-mockOutputs[1].TChan:
+				if string(ts.Payload.Get(0)) != string(content[0]) {
+					t.Errorf("Wrong content returned %s != %s", ts.Payload.Get(0), content[0])
+				}
+			case <-mockOutputs[0].TChan:
+				t.Error("Received message in wrong order")
+				return
+			case <-mockOutputs[2].TChan:
+				t.Error("Received message in wrong order")
+				return
+			case <-time.After(time.Second):
+				t.Errorf("Timed out waiting for broker propagate")
+				return
+			}
+
+			select {
+			case ts.ResponseChan <- types.NewSimpleResponse(nil):
+			case <-time.After(time.Second):
+				t.Errorf("Timed out responding to broker")
+				return
+			}
+		}()
+
+		select {
+		case res := <-resChan:
+			if res.Error() != nil {
+				t.Error(res.Error())
+			}
+		case <-time.After(time.Second):
+			t.Errorf("Timed out responding to broker")
+			return
+		}
+	}
+
+	oTM.CloseAsync()
+	if err := oTM.WaitForClose(time.Second * 10); err != nil {
+		t.Error(err)
+	}
+}
+
+func TestTryAllFail(t *testing.T) {
+	outputs := []types.Output{}
+	mockOutputs := []*MockOutputType{
+		{},
+		{},
+		{},
+	}
+
+	for _, o := range mockOutputs {
+		outputs = append(outputs, o)
+	}
+
+	readChan := make(chan types.Transaction)
+	resChan := make(chan types.Response)
+
+	oTM, err := NewTry(outputs, metrics.DudType{})
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	if err = oTM.StartReceiving(readChan); err != nil {
+		t.Error(err)
+		return
+	}
+
+	for i := 0; i < 10; i++ {
+		content := [][]byte{[]byte(fmt.Sprintf("hello world %v", i))}
+		select {
+		case readChan <- types.NewTransaction(types.NewMessage(content), resChan):
+		case <-time.After(time.Second):
+			t.Errorf("Timed out waiting for broker send")
+			return
+		}
+
+		testErr := errors.New("test error")
+		go func() {
+			for j := 0; j < 3; j++ {
+				var ts types.Transaction
+				select {
+				case ts = <-mockOutputs[j%3].TChan:
+					if string(ts.Payload.Get(0)) != string(content[0]) {
+						t.Errorf("Wrong content returned %s != %s", ts.Payload.Get(0), content[0])
+					}
+				case <-mockOutputs[(j+1)%3].TChan:
+					t.Errorf("Received message in wrong order: %v != %v", j%3, (j+1)%3)
+					return
+				case <-mockOutputs[(j+2)%3].TChan:
+					t.Errorf("Received message in wrong order: %v != %v", j%3, (j+2)%3)
+					return
+				case <-time.After(time.Second):
+					t.Errorf("Timed out waiting for broker propagate")
+					return
+				}
+
+				select {
+				case ts.ResponseChan <- types.NewSimpleResponse(testErr):
+				case <-time.After(time.Second):
+					t.Errorf("Timed out responding to broker")
+					return
+				}
+			}
+		}()
+
+		select {
+		case res := <-resChan:
+			if exp, act := testErr, res.Error(); exp != act {
+				t.Errorf("Wrong error returned: %v != %v", act, exp)
+			}
+		case <-time.After(time.Second):
+			t.Errorf("Timed out responding to broker")
+			return
+		}
+	}
+
+	oTM.CloseAsync()
+	if err := oTM.WaitForClose(time.Second * 10); err != nil {
+		t.Error(err)
+	}
+}
+
+//------------------------------------------------------------------------------